tests/wpt/web-platform-tests/css/css-animations/parsing/animation-name-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Animations: getComputedStyle().animationName</title>
<link rel="help" href="https://drafts.csswg.org/css-animations/#propdef-animation-name">
<meta name="assert" content="animation-name computed value is as specified.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<body>
<div id="target"></div>
<script>
test_computed_value("animation-name", 'none');

test_computed_value("animation-name", 'foo');
test_computed_value("animation-name", 'Both');
test_computed_value("animation-name", 'ease-in');
test_computed_value("animation-name", 'infinite');
test_computed_value("animation-name", 'paused');
test_computed_value("animation-name", 'first, second, third');

// TODO: Test strings, after https://github.com/w3c/csswg-drafts/issues/2435
// is resolved.
// Examples that need testing either here or in animation-name-invalid.html :
// '"Initial"', '"initial"', '"None"', '"Default"', '" x "', "1", '" "', '""',
// '"multi word string"', '"---\\22---"'
</script>
</body>
</html>
